+++ crates/proving-service-client/src/proving_service/tx_prover.rs
@@ -69,7 +69,7 @@ impl RemoteTransactionProver {
         let new_client = {
             ApiClient::connect(self.endpoint.clone())
                 .await
-                .map_err(|_| RemoteProverError::ConnectionFailed(self.endpoint.to_string()))?
+                .map_err(|err| RemoteProverError::ConnectionFailed(err.into()))?
         };
 
         *client = Some(new_client);