+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 
 - Added support for gRPC server side reflection to all components (#949).
+- Added support for TLS to `miden-proving-service-client` (#TODO).
 
 ### Changes